fix(web): address CodeRabbit feedback on debug-log hooks
- file-browser-hooks: drop sessionId from the init-effect deps so same-environment
  tab switches don't reset the tree (loadTree is intentionally stable via sessionIdRef)
- use-session-agentctl: include sessionId in the transition dedupe snapshot so the
  first transition after a session switch isn't suppressed
- file-browser-actions: extract scroll-persistence, child-loading, expand-toggle and
  file-open helpers into a sibling file to keep file-browser-hooks under the 600-line
  ESLint cap after the debug logging additions
- fix dangling `candidates` reference in applyFileChanges debug log
- switch IS_DEBUG tests to vi.stubEnv (NODE_ENV is readonly in TS 5.x)
